Refactor navbar component to use responsive sizing
