+++ next-app/src/app/page.js
@@ -5,7 +5,7 @@ import Projects from '@/components/Projects-Page/Projects';
 import Skills from '@/components/Skills-Page/Skills';
 import Awards from '@/components/Awards-Page/Awards';
 import ParticleBackground from '@/components/ParticleBackground';
-import NavBar from '@/components/NavBar';
+import NavBar from '@/components/Navbars/NavBar';
 import Footer from '@/components/Footer/Footer';
 
 export const metadata = {
@@ -33,9 +33,10 @@ const MainPage = () => {
           <About />
         </section>
 
-        {/* <section id="skills">
+        <section id="skills">
           <Skills />
         </section>
+        {/* 
 
         <section id="awards">
           <Awards />
