+++ next-app/src/components/Navbars/NavBar.js
@@ -3,12 +3,12 @@
 import { Link as ScrollLink } from 'react-scroll';
 import { Toolbar, Box, AppBar } from '@mui/material';
 import { useState } from 'react';
+import './ScrollLink.css';
 
 // Inline style for both active and inactive links
 const linkStyles = {
   margin: '0 16px',
   fontWeight: 700,
-  fontSize: '1.2rem',
   textDecoration: 'none',
   cursor: 'pointer',
   fontFamily: 'Montserrat, Arial, sans-serif',
@@ -23,6 +23,7 @@ const FormattedLink = ({ page, active, setActivePage }) => {
       offset={-70}
       duration={500}
       onSetActive={() => setActivePage(page)}
+      className="scroll-link"
       style={{
         ...linkStyles,
         color: active ? 'white' : 'gray',
@@ -46,10 +47,11 @@ const NavBar = () => {
         zIndex: 999,
         height: '7vh',
         backgroundColor: '#333',
+        height: '100%',
       }}
     >
-      <Toolbar sx={{ display: 'flex', alignItems: 'center' }}>
-        <Box sx={{ display: 'flex', alignItems: 'center', marginLeft: 'auto' }}>
+      <Toolbar sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center' }}>
+        <Box sx={{ display: 'flex', alignItems: 'center', marginLeft: { xs: '0', sm: 'auto' } }}>
           {pages.map((page) => (
             <FormattedLink key={page} page={page} active={activePage === page} setActivePage={setActivePage} />
           ))}
